escape-orcs-2: Fix character movements

+++ examples/escape-orcs-2/character.rs
@@ -1,7 +1,7 @@
 use crate::game_state::Movement;
 
 pub(crate) struct Character {
-    pub initial_position: (f32, f32),
+    pub initial_position: (i32, i32),
     pub animation_time: f32,
     pub movement: Movement,
 }
